requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.7.0
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
ruff==0.0.267
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
